+++ Packages/BrandonUtils/Standalone/Clerical/BPath.cs
@@ -1,6 +1,8 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using System.Text.RegularExpressions;
 
 using BrandonUtils.Standalone.Collections;
 using BrandonUtils.Standalone.Optional;
@@ -16,34 +18,42 @@ namespace BrandonUtils.Standalone.Clerical {
     /// </summary>
     [PublicAPI]
     public static class BPath {
-        internal static readonly RegexGroup ExtensionGroup = new RegexGroup(nameof(ExtensionGroup), @"(\.[^.]+?)+$");
-        internal static          char[]     Separators     = Enum.GetValues(typeof(DirectorySeparator)).Cast<DirectorySeparator>().Select(DirectorySeparatorExtensions.ToChar).ToArray();
+        internal static readonly RegexGroup ExtensionGroup            = new RegexGroup(nameof(ExtensionGroup), @"(\.[^.]+?)+$");
+        internal static readonly char[]     Separators                = Enum.GetValues(typeof(DirectorySeparator)).Cast<DirectorySeparator>().Select(DirectorySeparatorExtensions.ToChar).ToArray();
+        internal static readonly Regex      DirectorySeparatorPattern = new Regex(@"[\\\/]");
 
-        public static Failable ValidatePath(string maybePath) {
-            Action action = () => _ = Path.GetFullPath(maybePath);
+        public static Failable ValidatePath([CanBeNull] string maybePath) {
+            Action action = () => _ = Path.GetFullPath(maybePath!);
             return action.Try();
         }
 
-        public static Failable ValidateFileName(string maybeFileName) {
+        public static Failable ValidateFileName([CanBeNull] string maybeFileName) {
             Action action = () => {
                 ValidateFileNameCharacters(maybeFileName);
-                _ = Path.GetFullPath(maybeFileName);
+                _ = Path.GetFullPath(maybeFileName!);
             };
             return action.Try();
         }
 
-        private static void ValidateFileNameCharacters(string maybeFileName) {
+        [ContractAnnotation("null => stop")]
+        private static void ValidateFileNameCharacters([CanBeNull] string maybeFileName) {
+            if (maybeFileName == null) {
+                throw new ArgumentNullException($"The string [{maybeFileName.ToString(Prettification.DefaultNullPlaceholder)}] wasn't a valid filename: it was blank!");
+            }
+
             if (maybeFileName.ContainsAny(Path.GetInvalidFileNameChars())) {
                 var badCharacters = maybeFileName.Intersect(Path.GetInvalidFileNameChars());
                 throw new ArgumentException($"The string [{maybeFileName}] isn't a valid filename: it contains the illegal characters {badCharacters.Prettify()}!");
             }
         }
 
-        public static bool IsValidPath(string maybePath) {
+        [ContractAnnotation("null => false")]
+        public static bool IsValidPath([CanBeNull] string maybePath) {
             return ValidatePath(maybePath).Failed == false;
         }
 
-        public static bool IsValidFileName(string maybeFileName) {
+        [ContractAnnotation("null => false")]
+        public static bool IsValidFileName([CanBeNull] string maybeFileName) {
             return ValidateFileName(maybeFileName).Failed == false;
         }
 
@@ -54,21 +64,39 @@ namespace BrandonUtils.Standalone.Clerical {
         /// <remarks>This uses the <see cref="ExtensionGroup"/> <see cref="RegexGroup"/> for matching.</remarks>
         /// <param name="path">a path or file name</param>
         /// <returns><b>all</b> of the extensions at the end of the <paramref name="path"/></returns>
-        public static string[] GetExtensions(string path) {
+        [CanBeNull]
+        [ContractAnnotation("path:null => null")]
+        public static string[] GetExtensions([CanBeNull] string path) {
+            if (path == null) {
+                return null;
+            }
+
             return GetFullExtension(path)
                    .Split(new[] { '.' }, StringSplitOptions.RemoveEmptyEntries)
                    .Select(it => it.PrependIfMissing("."))
                    .ToArray();
         }
 
-        public static string GetFullExtension(string path) {
+        [CanBeNull]
+        [ContractAnnotation("path:null => null")]
+        public static string GetFullExtension([CanBeNull] string path) {
+            if (path == null) {
+                return null;
+            }
+
             return Path.GetFileName(path)
                        .Match(ExtensionGroup.Regex)
                        .Groups[ExtensionGroup.GroupName]
                        .Value;
         }
 
-        public static string GetFileNameWithoutExtensions(string path) {
+        [CanBeNull]
+        [ContractAnnotation("path:null => null")]
+        public static string GetFileNameWithoutExtensions([CanBeNull] string path) {
+            if (path == null) {
+                return null;
+            }
+
             var fileName    = Path.GetFileName(path);
             var firstPeriod = fileName.IndexOf(".", StringComparison.Ordinal);
             return firstPeriod < 0 ? fileName : fileName.Substring(0, firstPeriod);
@@ -76,31 +104,49 @@ namespace BrandonUtils.Standalone.Clerical {
 
         [ContractAnnotation("null => false")]
         public static bool EndsWithSeparator([CanBeNull] string path) {
-            path = path?.TrimEnd();
-            return path != null && Separators.Any(it => path.EndsWith(it.ToString()));
+            return path.EndsWith(DirectorySeparatorPattern);
         }
 
         [ContractAnnotation("null => false")]
         public static bool StartsWithSeparator([CanBeNull] string path) {
-            path = path?.TrimStart();
-            return path != null && Separators.Any(it => path.StartsWith(it.ToString()));
+            return path.StartsWith(DirectorySeparatorPattern);
         }
 
-        public static string EnsureTrailingSeparator(string path, DirectorySeparator separator = DirectorySeparator.Universal) {
-            path = NormalizeSeparators(path, separator);
-            return EndsWithSeparator(path) ? path : path.TrimEnd().Suffix(separator.ToChar().ToString());
+        [NotNull]
+        public static string EnsureTrailingSeparator([CanBeNull] string path, DirectorySeparator separator = DirectorySeparator.Universal) {
+            return path.TrimEnd(separator.ToCharString()).Suffix(separator.ToCharString());
         }
 
-        public static string StripLeadingSeparator(string path, DirectorySeparator separator = DirectorySeparator.Universal) {
-            path = NormalizeSeparators(path, separator);
-            return StartsWithSeparator(path) ? path.TrimStart().Prefix(separator.ToChar().ToString()) : path;
+        [NotNull]
+        public static string StripLeadingSeparator([CanBeNull] string path, DirectorySeparator separator = DirectorySeparator.Universal) {
+            return NormalizeSeparators(path?.TrimStart(separator.ToCharString()), separator);
         }
 
-        public static string NormalizeSeparators(string path, DirectorySeparator separator = DirectorySeparator.Universal) {
-            return Separators.Aggregate(
-                path.Trim(),
-                (current, c) => current.Replace(c, separator.ToChar())
-            );
+        [NotNull]
+        public static string NormalizeSeparators([CanBeNull] string path, DirectorySeparator separator = DirectorySeparator.Universal) {
+            return path.IsEmpty() ? "" : DirectorySeparatorPattern.Replace(path, separator.ToCharString());
+        }
+
+        [NotNull]
+        public static string JoinPath(
+            [CanBeNull] string parent,
+            [CanBeNull] string child,
+            DirectorySeparator separator = DirectorySeparator.Universal
+        ) {
+            parent = parent?.Trim().TrimEnd(DirectorySeparatorPattern);
+            child  = child?.Trim().TrimStart(DirectorySeparatorPattern);
+            var path = parent.JoinWith(child, separator.ToCharString());
+            return NormalizeSeparators(path, separator);
+        }
+
+        [NotNull]
+        public static string JoinPath([CanBeNull, ItemCanBeNull] params string[] parts) {
+            return JoinPath(parts, default(DirectorySeparator));
+        }
+
+        [NotNull]
+        public static string JoinPath([CanBeNull, ItemCanBeNull] IEnumerable<string> parts, DirectorySeparator separator = DirectorySeparator.Universal) {
+            return parts?.Aggregate((pathSoFar, nextPart) => JoinPath(pathSoFar, nextPart, separator)) ?? "";
         }
     }
 }